Refactor selection logic
